+++ src/lumon/solve/__init__.py
@@ -1,7 +1,11 @@
+from lumon.solve.ilp import SolverError, solve_budgeted_max_coverage, solve_min_cost_cover
 from lumon.solve.solution import InfeasibleError, OptimalityGuarantee, Solution
 
 __all__ = [
     "InfeasibleError",
     "OptimalityGuarantee",
     "Solution",
+    "SolverError",
+    "solve_budgeted_max_coverage",
+    "solve_min_cost_cover",
 ]